make separate folders for apps
