+++ src/js/pages/attendee/Event.js
