+++ src/js/pages/attendee/Events.js
