+++ src/js/pages/attendee/Info.js
