+++ src/js/pages/attendee/NewsFeed.js
