+++ src/js/pages/attendee/People.js
